requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
click>=8.0.0
rich>=10.0.0